q2

+++ b.py
@@ -1,1 +1,1 @@
-while 1:a,b=raw_input().split();print float(a)/10**int(b)
+while 1:a,b=raw_input().split();print int(a)/10.0**int(b)
